requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.3
Flask-Dance==6.2.0
requests
psycopg2-binary